GF#20968 Incomplete sentences in section 10.3.2 of populate.xml
Ended the sentence with "other answers within the form" based on context-
"It's possible that not all queries will resolve successfully - or that some context variables might not be populated.  In this situation, the system performing the population SHOULD make best efforts and populate whatever elements it can that are not dependent on other answers within the form."

Added a period at the end of "It's possible that the same data might be queried multiple times within different contexts when filling out the form.  Systems are free to notice this and use the data already in memory rather than re-querying the data."