GF#21355 , Added proposed language to 8.1.1 minValue & maxValue sections
Added the following language into two places in the behavior.xml file (under minValue and maxValue sub-headers in the 8.1.1 Value Constraints sectoin).

"The minValue SHALL be less than or equal to the <a href="#maxValue">maxValue</a>, if both are specified. If min and/or max are calculated values, care should be taken to ensure that they can't ever conflict.  I.e. It should never be possible for calculated max to be less than calculated min.  If this does occur, systems SHALL prohibit a value from being specified for the element for so long as the conflict exists.  If the question is required,  this may mean that it is not possible to mark the QuestionnaireResponse as "completed"."